refactor: accomoodation dto 생성
