+++ src/main/java/codesquad/team05/web/accommodation/dto/request/AccommodationUpdate.java
@@ -1,4 +1,4 @@
-package codesquad.team05.web.dto.request.accommodation;
+package codesquad.team05.web.accommodation.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotEmpty;
