+++ src/main/java/codesquad/team05/web/accommodation/dto/response/AccommodationResponse.java
@@ -1,7 +1,10 @@
-package codesquad.team05.web.dto.response.accommodation;
+package codesquad.team05.web.accommodation.dto.response;
 
+import codesquad.team05.web.picture.dto.response.PictureDto;
 import lombok.Data;
+import org.springframework.format.annotation.DateTimeFormat;
 
+import java.time.LocalDate;
 import java.util.List;
 
 @Data
@@ -17,5 +20,9 @@ public class AccommodationResponse {
     private String description;
     private String amenity;
     private List<PictureDto> pictures;
+    private boolean isOnSale;
+    private Double discountRate;
+    @DateTimeFormat(pattern = "yyyyMMdd")
+    private LocalDate endDate;
 
 }
